requirements.txt
# requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.52
Flask-Login==0.6.3
Werkzeug==2.3.7
python-dotenv==1.0.1
Flask-WTF==1.2.1
email-validator==2.2.0
Pillow==10.3.0
opencv-python==4.10.0.84
pyzbar==0.1.9
numpy==2.0.2
Flask-Migrate==4.0.7
pyOpenSSL==24.2.1